+++ src/components/Page/styles.js
@@ -1,13 +1,16 @@
 import { StyleSheet } from 'react-native';
 
-import styleConstants from '../../styleConstants';
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     alignSelf: 'stretch',
-    alignItems: 'center',
     position: 'relative',
+  },
+  testing: {
+    flex: 0,
+    width: 320,
+    height: 480,
+    alignSelf: 'center',
   },
 });